requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
asyncio==3.4.3
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
fastembed==0.3.4
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
loguru==0.7.2
mmh3==4.1.0
mpmath==1.3.0
nats-py==2.8.0
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.19.0
packaging==24.1
pillow==10.4.0
protobuf==5.27.3
pydantic==2.8.2
pydantic-core==2.20.1
pystemmer==2.2.0.1
pyyaml==6.0.2
redis==5.0.8
requests==2.32.3
snowballstemmer==2.2.0
sympy==1.13.2
tokenizers==0.20.0
tqdm==4.66.5
typing==3.7.4.3
typing-extensions==4.12.2
urllib3==2.2.2
